+++ src/python/DAS/utils/das_config.py
@@ -114,8 +114,10 @@ DASOption('web_server', 'dbs_daemon', 'bool', False),
 # Onhold daemon used to process onhold request, those requests
 # are made by DAS server against DDoS clients
 DASOption('web_server', 'onhold_daemon', 'bool', False),
-# Threshold for DDoS clients
+# Threshold for DAS normal clients
 DASOption('web_server', 'hot_threshold', 'int', 100),
+# Threshold for DAS super clients
+DASOption('web_server', 'super_hot_threshold', 'int', 5000),
 # onhold_daemon option
 DASOption('web_server', 'onhold_daemon', 'bool', False),
 # DBS dataset daemon update interval in seconds